requirements.txt
beautifulsoup4==4.11.1
html_to_json==2.0.0
pandas==1.3.3
python_dateutil==2.8.2
requests==2.26.0
selenium==3.141.0
streamlit==1.9.0
